requirements.txt
flask>=2.0
mysql-connector-python>=8.0.5
gunicorn>=20.1.0
#WTForms>=2.3.3